+++ doordeck-sdk/src/commonMain/kotlin/com/doordeck/multiplatform/sdk/model/responses/PlatformResponses.kt
@@ -37,9 +37,9 @@ internal sealed interface BasicAuthKeyResponse {
     val x5t: String?
     val x5t256: String?
     val x5c: List<String>?
-    val exp: Int?
-    val nbf: Int?
-    val iat: Int?
+    val exp: Long?
+    val nbf: Long?
+    val iat: Long?
 }
 
 @Serializable
@@ -55,9 +55,9 @@ internal data class BasicRsaKeyResponse(
     @SerialName("x5t#S256")
     override val x5t256: String? = null,
     override val x5c: List<String>? = null,
-    override val exp: Int? = null,
-    override val nbf: Int? = null,
-    override val iat: Int? = null,
+    override val exp: Long? = null,
+    override val nbf: Long? = null,
+    override val iat: Long? = null,
     val e: String,
     val n: String
 ): BasicAuthKeyResponse
@@ -75,9 +75,9 @@ internal data class BasicEcKeyResponse(
     @SerialName("x5t#S256")
     override val x5t256: String? = null,
     override val x5c: List<String>? = null,
-    override val exp: Int? = null,
-    override val nbf: Int? = null,
-    override val iat: Int? = null,
+    override val exp: Long? = null,
+    override val nbf: Long? = null,
+    override val iat: Long? = null,
     val crv: String,
     val x: String,
     val y: String
@@ -96,9 +96,9 @@ internal data class BasicEd25519KeyResponse(
     @SerialName("x5t#S256")
     override val x5t256: String? = null,
     override val x5c: List<String>? = null,
-    override val exp: Int? = null,
-    override val nbf: Int? = null,
-    override val iat: Int? = null,
+    override val exp: Long? = null,
+    override val nbf: Long? = null,
+    override val iat: Long? = null,
     val crv: String,
     val x: String
 ): BasicAuthKeyResponse
